pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-trader"
version = "0.1.0"
description = "Auto trader scaffold (SQLite + pydantic settings + logging)"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
  "pydantic>=2.5",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "requests>=2.31",
  "websockets>=12.0",
  "httpx>=0.27",
  "discord.py>=2.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
  "asgi-lifespan>=2.1",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["v2/tests"]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["v2*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "BLE"]
ignore = ["E501", "B008", "BLE001"]

[tool.ruff.lint.isort]
known-first-party = ["v2"]
